requirements.txt
lxml
beautifulsoup4
python-dotenv
bs4
python-dotenv
opencv-python
numpy
requests
grpcio
grpcio-tools
protobuf